requirements.txt
altair==5.3.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
Backtesting==0.3.3
beautifulsoup4==4.12.3
blinker==1.8.2
bokeh==3.5.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.3
decorator==5.1.1
executing==2.0.1
fastjsonschema==2.20.0
fonttools==4.53.1
frozendict==2.4.4
gevent==24.2.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
html5lib==1.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
logzero==1.7.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multitasking==0.0.11
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyotp==2.9.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
setuptools==72.1.0
six==1.16.0
smartapi-python==1.4.9
smmap==5.0.1
soupsieve==2.5
stack-data==0.6.3
streamlit==1.37.1
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket==0.2.1
xyzservices==2024.6.0
yfinance==0.2.41
zope.event==5.0
zope.interface==6.4.post2
